+++ app/Text/ValveVKV.hs
@@ -1,5 +1,5 @@
-module Text.ValveVKV(parseValveVKV, parseToVKV, fromValveVKV, (.:), (^:), unpair, ValveVKV,
-    ValveKeyValueEntry(KVObject, KVInt, KVString), Pair (Pair), Context) where
+module Text.ValveVKV(parseValveVKV, parseToVKV, fromValveVKV, (.:), (^:), ValveVKV,
+    ValveKeyValueEntry(KVObject, KVInt, KVString), Pair (Pair), unpair, Context, vkvParser) where
 -- Library for processing Valve's value keyvalue format. The main function you will wish to use is parseValveVKV. To convert it into your own type, you
 -- will need to write a 'ValveVKV' instance for it.
 
@@ -9,11 +9,11 @@ import Text.ValveVKV.Class
 import Data.Maybe (mapMaybe)
 
 -- | The main function you will be using. Turns the ValveVKV string into a type that has the 'ValveVKV' typeclass.
-parseValveVKV :: ValveVKV a => String -> Maybe a
+parseValveVKV :: ValveVKV a => String -> Either String a
 parseValveVKV input =
     let parseRes = parse vkvParser "" input in
     case parseRes of
-        Left _ -> Nothing
+        Left s -> Left (show s)
         Right a ->
             let topObj = KVObject (Pair "top" a) in
             fromValveVKV topObj topObj